remove unwanted classes
