+++ DirectDebitSubmissionNightyJob/BaseFunction.cs
@@ -1,16 +1,14 @@
-using Amazon.XRay.Recorder.Core;
-using Amazon.XRay.Recorder.Core.Strategies;
 using Amazon.XRay.Recorder.Handlers.AwsSdk;
 using DirectDebitSubmissionNightyJob.Infrastructure;
 using Hackney.Core.Logging;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Reflection;
 using System.Text.Json;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace DirectDebitSubmissionNightyJob
 {
