--- DirectDebitSubmissionNightyJob/Boundary/Request/DirectDebitExportRequest.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Collections.Generic;
-
-namespace DirectDebitSubmissionNightyJob.Boundary.Request
-{
-    public class DirectDebitExportRequest
-    {
-        public string FileType { get; set; }
-        public DateTime? Date { get; set; }
-        public List<Guid> SelectedDirectDebitLists { get; set; }
-    }
-}
