--- DirectDebitSubmissionNightyJob/Boundary/Response/DirectDebitMaintenanceResponseObjectList.cs
@@ -1,9 +1,0 @@
-using System.Collections.Generic;
-
-namespace DirectDebitSubmissionNightyJob.Boundary.Response
-{
-    public class DirectDebitMaintenanceResponseObjectList
-    {
-        public List<DirectDebitMaintenanceResponseObject> ResponseObjects { get; set; }
-    }
-}
