+++ DirectDebitSubmissionNightyJob/Gateway/Interfaces/ITransactionApiService.cs
@@ -4,7 +4,7 @@ using System.Text;
 using System.Threading.Tasks;
 using DirectDebitSubmissionNightyJob.Boundary.Request;
 
-namespace DirectDebitSubmissionNightyJob.Services.Interfaces
+namespace DirectDebitSubmissionNightyJob.Gateway.Interfaces
 {
     public interface ITransactionApiService
     {
