--- DirectDebitSubmissionNightyJob/Services/Interfaces/IAccountApiService.cs
@@ -1,14 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-using DirectDebitSubmissionNightyJob.Boundary.Request.Account;
-using DirectDebitSubmissionNightyJob.Boundary.Response.Account;
-
-namespace DirectDebitSubmissionNightyJob.Services.Interfaces
-{
-    public interface IAccountApiService
-    {
-        Task UpdateRentAccountBalance(AccountUpdateRequest request);
-    }
-}
