--- DirectDebitSubmissionNightyJob/Services/Interfaces/IHousingSearchApiService.cs
@@ -1,13 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-using DirectDebitSubmissionNightyJob.Boundary.Response.Account;
-
-namespace DirectDebitSubmissionNightyJob.Services.Interfaces
-{
-    public interface IHousingSearchApiService
-    {
-        Task<AccountResponse> GetRentAccountByPrn(string paymentReference);
-    }
-}
