--- DirectDebitSubmissionNightyJob/Services/Interfaces/ITenureApiService.cs
@@ -1,13 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-using DirectDebitSubmissionNightyJob.Boundary.Response;
-
-namespace DirectDebitSubmissionNightyJob.Services.Interfaces
-{
-    public interface ITenureApiService
-    {
-        Task<TenureResponse> GetTenureInformation(Guid targetId);
-    }
-}
